HFTS_Register/requirements.txt
flask
pillow
pyodbc
numpy
deepface
tf-keras
azure-storage-blob
